+++ src/register.tsx
@@ -8,8 +8,6 @@ import Panel from './components/Panel/Panel';
 import SettingsProvider from './contexts/SettingsProvider';
 
 addons.register('storybook-addon-theme-playground', api => {
-  // Also need to set a unique name to the panel.
-
   addons.addPanel('storybook-addon-theme-playground/panel', {
     title: 'Theme Playground',
     render: ({ active, key }) => (