+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0]- 2020-01-08
+
+### Added
+
+- Ability to hide specific theme values or objects
+- Added `debounceRate` to config
+- Loading state for debounced theme updates
+- HSL and default css colors for color value check
+- Reset button
+
+### Changed
+
+- State handling inside settings context
+- Memoized values on `RadioGroup` and settings components
+
 ## [1.1.0]- 2020-01-05
 
 ### Added
@@ -34,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.1.0...develop
+[unreleased]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.2.0...develop
+[1.2.0]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.0.0...v1.0.1